[FIX] sap.m.ActionSelect: Fixed incorrect scroll behavior
Problem: When there is a scroll in the flex area of SimpleFixFlex and we
are navigating in the list items, after going outside of the visible
part of the list, the scroll is not responding. Resulting in a selected
offscreen item.

Root Cause: Since the introduction of the SimpleFixFlex private control,
the list is inside its flex area. The scroll responds only for the control
in the flex area (the list itself), but the ActionSelect's buttons are
not in it, but in the Picker. This results in incorrect scroll
behavior of the flex area (parent of the list items).

Solution: We enable the flex area of the SimpleFixFlex to accept more
than one control and put the ActionSelect's buttons inside it.

Change-Id: I1eae18b3ac9a2dda87e166a99f6293e8f62e79a1
BCP: 1980194996