[FIX] sap.m.SinglePlanningCalendar: View's title is now always propagated to the view switch
The SinglePlanningCalendar gets notified when the title
property of a containing view gets changed.

Example of a use case: When there is а title of a view,
which is bound to a model, the SinglePlanningCalendar
now sets it to the corresponding SegmentedButtonItem
from the views switch in the PlanningCalendarHeader.
This is handled via a usage of ManagedObjectObserver.

BCP: 0020751295 0000357152 2019

Change-Id: I4c454467a53f651ee7d66608b7607d1817ec42db
CR-Id: 002075125800003782132019