[FIX] sap.f.FlexibleColumnLayout: Mid column does not disappear now
Problem: In some cases, on back navigation the middle column of
FlexibleColumnLayout disappears. This happens because the resize handler
function is called two times for each column, and the first time it has
0px width. Once, the 'sapFFCLColumnHidden' is added by
_adjustColumnDisplay function, it does not getting removed when it is
called second time with different value.

Solution: The _adjustColumnDisplay function is adjusted, so that now if
the width is different than 0, the 'sapFFCLColumnHidden' is getting
removed.

Change-Id: I4906f23f339f42f2a430f5979108a2b2ebb7ab8f
BCP: 1970341860