[INTERNAL] Table: Added row modes that provide additional render options
- Added hidden "rowMode" aggregation
- Added Fixed-/Interactive-/AutoRowMode to replace VisibleRowCountMode
  Fixed/Interactive/Auto (sap.ui.table.rowmodes.*)

New features:
- FixedRowMode:
  + Hide empty rows and show a placeholder for them at the bottom of the
    table.
- InteractiveRowMode:
  + Allow the user to shrink the table to the configured minimum row. In
    the old "Interactive" VisibleRowCountMode, the user was limited to
    the number of fixed rows +1. This behavior can still be enabled by
    setting the "minRowCount" property accordingly.
- AutoRowMode:
  + Hide empty rows and show a placeholder for them at the bottom of the
    table.
  + Set a maximum row count.

Fixes:
- AutoRowMode
  + If the fixed top row count is bigger than the minimum row count, the
    scrollbar would disappear if the visible rows decrease to <= fixed
    top row count. This is also fixed for VisibleRowCountMode=Auto.

The row mode classes are currently private and used by default
internally by the table. Every mode is able to operate in a "legacy
mode", to still support the old table API (e.g. visibleRowCountMode,
visibleRowCount, ...) and the old behavior.
If no row mode is set, which is currently only allowed for the MDCTable,
the table instantiates the correct row mode according to the
"visibleRowCountMode" property and makes it operate in legacy mode.

All row modes, if not in legacy mode, enforce fixed row heights. The
height is configurable with a "rowContentHeight" property, otherwise a
default row height is applied by the table.
Row height handling has also been moved to the row modes, because the
table's "rowHeight" property is tightly coupled to the old
"visibleRowCountMode" property with different behaviors per mode. A
change in behavior might justify a clear API separation. However, a
final decision has not yet been made.

Properties that are to be deprecated when row modes are made public:
- visibleRowCountMode
- visibleRowCount
- fixedRowCount,
- fixedBottomRowCount
- minAutoRowCount
(- rowHeight)?

Change-Id: Ia8833f11ed1d6fe8f57eb5a25779d18728ffcd4b
JIRA: CPOUIFTEAMB-1369