[INTERNAL] ItemNavigation: Recalculate focused index when items are set
 - ItemNavigation internally stores focused index to detect where to
move the focus when navigation keys are pressed and this value is
updated only on focusin.
 - In traditional string-based rendering focusin event was always
hitting after rendering if the focus was one of the navigation items.
With our new rendering engine DOM elements are not thrown away but only
patched. In this case if focused element is not changed then the browser
does not fire focusin event.
 - This change introduces a focused index recalculation after new items
are set.

Change-Id: I9e9ac03ba7d54463063feffadcf3d4e856acf44a