[INTERNAL] sap.f.GridContainer: Columns breathing is improved
Add minColumnSize and maxColumnSize properties to allow breathing.

Listen for the viewport resize, together with the grid resize.
Grid resize is not enough when the grid is centered and columns are limited.

Add different CSS class to the GridContainer for the different layouts
XL, L, M, S.
This is consistent with CSSGrid and allows more customization, if
needed.

Fixed the firing of the 'layoutChange' event.

Added example with containerQuery=true.

JIRA: BGSOFUIRODOPI-2081

Change-Id: I68565509f4b38a7159d9a717aaeb0727b73b86d9