[FIX] Demo Kit: Improved navigation from Samples to API Reference
Problem:
If the sample belongs to more than one entity and the user tries to
navigate using the  "API Reference" button, it doensn't navigate to
the correct entity.

Fix:
We are taking the context (control which sample belongs to)
for the entity from the router path and navigate correctly
(to correct control API Reference).

Change-Id: I32e2bc0130ce164a9cade06ab0c37415bd261c6f
Fixes: https://github.com/SAP/openui5/issues/2518
BCP: 1980281215