[FEATURE] Theme Origin Whitelist via <meta> tag
When providing the "sap-theme" URL parameter, a theme and a themeRoot
can be specified. Due to security reasons absolute URLs were not
allowed.
With this change allowed origins can be defined in a <meta> tag:

<meta name="sap-allowedThemeOrigins" content="https://example.com">

The "content" may contain multiple origins separated with comma.
A wildcard (*) can also be used to allow any origins.
However this should only be used in combination with additional security
mechanisms such as CSP style-src directives.

Change-Id: Ibf6ddd2c09cfbc8de84d87c50ee16d4992245ff8