[INTERNAL][FIX] sap.ui.layout.Grid: Fixed failing assertion for the renderer
Problem:
A certain line of code in the renderer was causing assertion exceptions
from the renderer API.
The argument for the "class" method of the renderer is supposed to have
only one css class/word.
Previously there was an option to pass a string which contains more than
one classes/words in it like "class1 class2", which was causing the
exceptions.

Solution:
We split the string ensuring that even if there is more than one
word/class, we will pass them one by one.

Change-Id: Ifc1645d4f1978164a1bf4819b077432ad15e8768
BCP: 1970359634
Fixes: https://github.com/SAP/openui5/issues/2599