[INTERNAL] sap.m.SimpleFixFlex: Max height of Fix area removed
Problem: If we have more lines of text in the Fix area than there
are items to select from the Flex area, there is an unnecessary scroll
in the Fix area.

Root Cause: The problem is caused from the fact that the Fix area
has a 50% max-height in its CSS. That causes the Fix area to have a
scrollbar, if its height is bigger than 50% of the whole SimpleFixFlex.

Solution: We remove the 50% CSS height restriction, which eliminates the
option for a scrollbar in the Fix area.

Additionally we added a warning, which indicates the App developer it's
not recommended to use more than 200 characters in text of the Fix area.

Change-Id: Ibb64082d3c10f7c5968f169d93dc7b9584d3349b
JIRA: BGSOFUIPIRIN-4077