[FEATURE] sap.f.Avatar: Introduced ability to display border
The new 'showBorder' boolean property of sap.f.Avatar now enables the
app developer to decide whether the control has a border or not.

The border is implemented with box-shadow according to the new Visual
Specification for Enhanced Avatar control. As in HCB and HCW themes the
Avatar already has a solid border, it is only applied for the other themes.

Change-Id: I42dd40ef799fd83cc28bf399537cf10459677f85
JIRA: BGSOFUIPIRIN-4062