[FIX] sap.m.Select: Focusout firefox behavior fixed
BCP: 1970364518

Problem: After a qUnit test start failing on Firefox only, it turned out
we always return from _handleFocuOut in rendering phase, where the patch
for Firefox was applied(onBeforeRendering).

Solution: The conclusion is, that Firefox now handles the focusout event
correctly, so there is no need of the patch anymore and the qUnit test
in question passes now.

Change-Id: I26b51872ca26d86870ed5ad47de7ae783653001d