[FIX] sap.m.OverflowToolbar: Changing visibility of content is fixed
Problem: When the visibility of a Control, placed in OverflowToolbar is
changed, and in the same JavaScript tick the _handeResize function is
called, wrong value for Control's size is cached. This leads to
incorrect behavior upon rerendering afterwards.

Solution: If the visibility of a Control, placed in OverflowToolbar is
changed, we assure that after the rerendering its size will be cached
again with a correct value.

Change-Id: I9bdd27480e1c19661d708ea3dfbe47cc9badc220
BCP: 1970072029