[FIX] ODataTreeBindingAdapter: Prefer $count value as initial binding length
The getLength implementation of TreeBindingAdapter.js depends on
the magnitude value of the root node. The initial calculation of this
value is triggered by a getContexts() call, which internally calls
_buildTree().

However, listeners of the bindings dataReceived event might expect the
bindings length to be available synchronously after the event has been
fired for the first time. This is already the current behavior of the
ODataListBinding.

This fix prefers the already known child count of the hierarchies
virtual root node (internally the "null" node) in cases where its
magnitude is not (yet) known.

BCP: 1970312273
Change-Id: Ifde4dac39aae32362e427f1319e77b0388710b79