[FIX] sap.m.PlanningCalendar: Appointment selection now properly indicated
Screen readers read that an appointment is selected if it has
a reference to a hidden 'Selected' text. Until now however, such
reference was added on every click, regardless if it selected
or deselected an appointment. This is now fixed - the reference
is added on selection and removed on deselection.

Change-Id: I22b43b304fc29c00248fb306cb1bed8b8291931c
BCP: 1970332529