[INTERNAL][FIX] sap.f.SemanticPage: Fixed volatile check in RTA
Problem: There is a check in the designtime file of the SemanticPage,
where we are expecting a dom ref of the SemanticPage, but such might
not exist. This results in exception.

Solution: We explicitly check if SemanticPage and its dom ref exist,
before we continue our execution in the if-else block.

Change-Id: Idddbda5411fe076a8d4fc388a9226c8747959394
BCP: 1970382337