[INTERNAL] UIArea-RenderManager: Preserved area handling
 - Before the re-rendering, UIArea was moving all "to-be-preserved" DOM
nodes to the preserved area and the control was responsible to move the
control DOM from preserved area to the active DOM tree on after
rendering.
 - With this change, moving "to-be-preserved" control DOM nodes to the
preserved area is left to the control responsibility. Controls should do
this onBeforeRendering hook via RenderManager.preserveContent API.
 - When a control is removed from the control tree or when the visible
property of the control is set to false, the invalidation and
re-rendering will not be executed on child controls but only on the
parent.
In this case, since the rendering cycle of the child controls is not
going to be executed, moving
"to-be-preserved" DOM nodes to the preserved area will be taken care of
the RenderManager via UIArea call.
 - HTML and XMLView controls are refactored accordingly.
 - This change is important to keep the rendering related logic of a
control inside the control rendering cycle. This will let us to move on
the "skip rendering of non-invalidated controls" inspiration.


Change-Id: I6eaad22f2e31b961f492285120ef3d8f5bad4604