[FIX] sap.m.ScrollBar: Fixed incorrect rendering on 80% browser zoom
Problem: On 80% browser zoom, part of the original browser's scrollbar
is visible, under the fake one.

Solution: Removed unnecessary width of the outer div and a negative
margin of the inner one.

Additionally, the sap.m.ScrollBar is now aligned with the new rendering
syntax.

Also the qUnit of the sap.m.ScrollBar was added to the qUnit list of
the sap.m library.

Change-Id: I94ba54cdf3c1f8d7a4f2eec1adf29a290a49a4bb
BCP: 1970285385