[INTERNAL] sap.ui.model.odata.v4.lib._Requestor: session lost w/o "Sap-Err-Id"
Do not rely on the server sending "Sap-Err-Id: ICMENOSESSION" if the
session is lost, because there are (exceptional) situations where this
header is not sent.

Always fire the session timeout event if the request contained a
Context-ID and the error response doesn't. Ignore the "Sap-Err-Id"
completely. A session only ends regularly with a successful response.
It is surely lost if the server is asked to process a request in it,
runs into an error and does not return its ID any more.

Change-Id: I65568614a66dc5a62da4f5b7dd02e34c5aa966f6
BCP: 1970366481
BCP: 1970366483
(cherry picked from commit 37f9869099212760921ba96178880b59cddb040b)