[FIX] ListItemBase: range selection fix
When the keydown/keyup event is triggered on the "MultiSelect" checkbox
control, the ListBase was not informed correctly to initialize/terminate
the range selection mode. Hence adapting the keydown/keyup events in the
ListItemBase such that the range selection mode is initialized/terminated
correctly also when the event is triggered on the "MultiSelect" checkbox
control.

Change-Id: I2b2248bed5831a8e455541e2fa0d62ee2eba4ee9