[INTERNAL] Table: Avoid adding legacy row modes to dependents
If the legacy row mode is added to the dependents aggregation, it is
also cloned. The cloned table would then create a new legacy row mode on
init and also add it to the dependents aggregation. To avoid memory
leaks and other potential issues, the legacy row mode is no longer added
to dependents.
Also some row mode internal API are changed.

Change-Id: I78f59182709b07f0c9c94c52018dfdb9aa1bd99d