[FIX] sap.m.Input: Selection in tabular suggestions fixed
Issue:
When autocomplete is turned on and there are suggestions.
Starting to type opens the suggestions popover. One of the
items is automatically selected and appears selected in
the table with a blue highlight color. Click on that item
will not close the suggestions popover and it will not change
the selection row of the table.

Solution:
We now check if the itemPress occured on an item which is
already selected. If this is the case we set the selectionRow
of the table.

NOTE: This does NOT happen in list suggestions as we attach
directly to itemPress. NOT sure what would be the consequence
of fully changing from selectionChange to itemPress.

BCP: 1970396382
Change-Id: I7412d93b0ab1e1435d8c076db7ca2a00b10a038f