[FIX] sap.m.TextArea: Sync properties growing + width + cols
As we're using a hidden container when growing="true",
it could not adjust itself out of the box when cols property is provided.
Further more, cols is a very special TextArea property which is defined
as "width of the text area (in average character width)" and is very hard
to be calculated. So, we're trying to mimic it by using half of the
default character width (0.5rem) as 1 col;

BCP: 1970303138

Change-Id: I46073bedd660983aabba1b70aaf15a54aaf48802