[INTERNAL] sap.ui.odata.v4.ODataModel: reduce even w/ different update group ID
When the property binding's update group ID differs from the one of the
binding with the cache, we did not reduce because a PATCH for this
binding would occur in the wrong update group.

However it is easily possible to ensure that the binding uses the update
group ID of its parent binding instead of letting the binding with the
cache decide it. Then it is no problem to reduce in this case, too.

PS1: red integration test & POC
PS2: TDD URE/SSC
PS3-5: review comments

Change-Id: Ie17a069a0026969a89bd020301ea4c4a823f598b
JIRA: CPOUI5UISERVICESV3-1942