[INTERNAL] Table: Deactivate table selection API if selection plugin set
If a selection plugin is applied to the table, the table's selection API
must not be used. The API of the plugin should be used instead. This is
already documented.
The reason is that we cannot guarantee functional correctness of table
API for selection plugins that might be added in future, especially not
for custom selection plugins. Therefore, a clear separation of APIs must
be ensured.

Furthermore, legacy plugins (and rowModes) are added in #applySettings.
This should avoid unnecessary creation and immediate destruction of
objects if they are already provided as part of the control settings.

Change-Id: I1b5a901b01c68ab018c5f0444b94bfbb2f5609fc