[FIX] sap.ui.core.hyphenation.Hyphenation: Out of memory errors fixed
Issue:
Initializing languages in hyphenation too many times results in
out of memory exception. Which sometimes causes side effects in
applications which uses it.

Solution:
Additional exception handling is added to ensure such errors dont
break applications.
We try to reduce the number of initializations as much as possible.

Reproducible with snippix: 66059

CR-Id: 002075125800004638052019
BCP: 002075129500003693692019
Change-Id: Ie8337ef37c66738b17ed832a53d7caa36a68bc17