[INTERNAL] Table: Delete horizontal scroll on column resizing qunit test
If a column is resized outside the table's viewport, there was a focusin
event after rendering to which the browser and the table reacted and
scrolled the resized column (which has the focus) into the viewport.
After migrating to the new DOM patching rendering, there is no focusin
after rendering. As this functionality now doesn't exist anymore and is
not considered to be essential, the corresponding tests are removed.

Change-Id: If41bd2f3bd829fdd23edf90deffa8a650c339c41