[INTERNAL] Table: fix styles for Fiori 3 design
The height in the cozy mode was incorrect due to the additional paddings
that were applied to the cells. Overwirting the style for Fiori 3 theme
cell height is according to the visual specs (2.75rem).
The background color of the GroupHeaderListItem was incorrect in the base
theme, hence correcting the background color according to the visual specs.
BCP: 1970419303
BCP: 1970388339

Change-Id: I23eaaf8f10a6b36f24c9346a915e603ce3a1f27e