[FIX] sap.uxap.HierarchicalSelect: Fixed incorrect alignment in FCL
Problem: When there is a sap.uxap.HierarchicalSelect in
sap.f.FlexibleColumnLayout, the visual representation of the
sap.uxap.HierarchicalSelect is incorrect.

Root Cause: A CSS selector which was applied only in
sap.f.FlexibleColumnLayout caused the height of the control to be
incorrect.

Solution: Removed the problematic CSS selector, since we don't need it
anyway.

Additionally:
Removed redundant "sapUiContainer" CSS classes from all of the
theme files of the sap.uxap.HierarchicalSelect.

Removed redundant Less code from the Fiori3 theme file, which was
generating the same CSS code for "cozy" and "compact". We ended up with
CSS for "cozy", which is sufficient for "compact", since by specification
they are the same.

Fixed a minor bug in all of the contrast theme files, where a left
border on the arrow-down button will show on a Phone devices.

Change-Id: Ib70f68e8199aa03f3ee87b3edb8852fa979f4f58
BCP: 1970422781