[INTERNAL] sap.m.Dialog: beginButton type change fixed for Quartz Light
No matter what type of 'beginButton' or 'endButton' the user gave,
they were turned to 'Emphasized' and 'Transparent'. This is not
wanted for sap.m.Dialog, but wanted for other types of dialogs,
where we have control over the buttons.

Now the following  controls have their 'beginButton' emphasized
for all themes:
- sap.m.SelectDialog
- sap.m.TableSelectDialog

Change-Id: Iee2bb98657e2e509ec65dedba4be0c51a2372977
JIRA: BGSOFUIRODOPI-2080