[FIX] sap.f.GridList: GridBoxLayout resizing optimizations
Issue:
When the GridList is configured to use GridBoxLayout
in combination with grouping and growing, the GridBoxLayout
is forced to manually equalize the heights of the items
resulting in performance issues when too many items are
rendered.

Solution:
As we cannot avoid the calculations I tried to optimize
the equalization function as much as I could.

BCP: 1930410831
Change-Id: Ib8f9460f4997ddfa03b371196c8916b360be8fd9