[FIX] sap.m.NotificationListItem: Expand/collapse improved
When jquery-ui was loaded on the page, the expand collapse
functionality of the notifications was entering an infinite loop.
This was caused by toggleClass's second parameter
wich was not strict boolean and this was triggering an animation.

Now this parameter is strict boolean
and the infinite expand/collapse does not happen.

Snippix: #94340

Change-Id: I2afd83078df782013fb967a996087de52173849a
BCP: 002075129500004216592019