[FIX] v2 binding: wrong preliminary context propagated
A preliminary context should be propagated only to bindings
using/creating preliminary contexts. As we reuse the contexts,
the same context could be used in diferent scenarios which
leads to problems in updating PropertyBindings.

Context is now set to preliminary = false on import of data so
this problem cannot happen anymore.

Change-Id: If3ba6a8033c9975d7abb198f41cfef127a6e8655
BCP: 1970331517