[FIX] ui5loader: avoid automatic chaining of promise exports
If an AMD module exports a promise, that promise should be given
to the original caller, not its result. This also implies that the
loader MUST NOT wait for the settlement of an exported promise before
resolving dependencies to the exporting module.

While this would be a nice and powerful feature, it has two significant
drawbacks:

1. it isn't in line with the AMD spec or other AMD implementations nor
with the ES6 module loading behavior. Both environments don't provide
such a capability. If the ui5loader would support it, it would become
much harder to generically migrate UI5 code to one of the two worlds.

2. when a module can only export things that are created during the
synchronous execution of its top-level module code, it is significantly
easier to determine the set of initially required dependencies at build
time. Assuming the use of AMD compliant APIs, there are simply no other
APIS than (sap.ui.)define itself that need to be understood. But when
waiting for the settlement of an exported promise would be part of the
loader behavior, any API could contribute to the set of eager
dependencies and determining the initial module set would become
impossible.

Change-Id: I1318a2557ca6210049cc18e68a16028fa3c9adc7
BCP: 1970448538