[FIX] sap.m.MultiComboBox: Layout corrected when icon is pressed
BCP: 1970441822
BCP: 1970432452

This reverts commit 0a524fbc0de370906a4a815d2d9347e4e59c2924 since the
change leads to broken layout on small screen sizes or small width of
the control. When a MultiComboBox has a small width and N-more is shown,
pressing the arrow down icon causes an incorrect layout and as a result
the dropdown will not open. What is more, the arrow icon moves around
the control.

Change-Id: I7a1836ab0291ddb6d29a3fc8373a4c57511d2af2