[INTERNAL] sap.ui.fl...apply.BrowserStorageConnector adjusted response
The not yet used BrowserConnector is prepared to provide the response
for the loadFlexData request in a different format. Instead of building
the variant section on its own the Connector.js will be in charge with a
following commit.
The accepted format in a flex response with:
{
	changes: [],
	variants: [],
	variantChanges: [],
	variantDependentControlChanges: [],
	variantManagementChanges: []
}
or an array of such objects. Each entry in the array is reflecting a
"layer" (derived by layer or connector like 'VENDOR' or 'CUSTOMER_BASE'
which are already actual 2 layers... one from the changes-bundle.json
and the back end.

Change-Id: Id1f33b029d9778f395b914aaf4f8b93935d19e38