[FIX] Remove sap.tnt.InfoLabel from sap.f.ShellBar
Before this fix the shellBar/Factory required sap.tnt.InfoLabel.
With this change the require was removed a) because the control
was not even used in the module and b) a hard reference from sap.f
to sap.tnt should not be made, as this will include the complete
library boostrap of sap.tnt and load the complete theming of
sap.tnt in every application using sap.f, even if no reference to
sap.tnt is needed.

Change-Id: Ibadc0de472a1169a1b73036d37048543ee9513d1