[INTERNAL][FIX] sap.ui.fl.transport.TransportSelection includes a reference
The publish behavior changed to handle descriptor changes not known by
the client. Therefore, a reference has to be sent and the server will
use the reference instead of changeIDs. The changeIDs still remain in
the request to ensure a proper handling on older back ends.
Also publishes changes for a given app version & layer.

Change-Id: I7d541b6a7913fc39897069adec84b084b7ca32dc
BCP: 1980322650
CR-Id: 002075125900004643732019