[INTERNAL][FIX] sap.ui.fl.transport.TransportSelection includes a reference
The publish behavior changed to handle descriptor changes not known by
the client. Therefore, a reference has to be sent and the server will
use the reference instead of changeIDs. The changeIDs still remain in
the request to ensure a proper handling on older back ends.
Publishes changes for a given app version & layer.

Change-Id: I0767af577a9f5d9a11f27f22e36d68141cdb466b
BCP: 1980322650
CR-Id: 002075125900004643742019