[INTERNAL] sap.ui.fl.apply._internal.Connector loads lists from connectors
This commit enables the sap.ui.fl.apply._internal.Connector.js to handle
also multiple responses from a single connector.
This is needed in case the data provided by the connector affects the
variant section and the response of the connector is already layered.
Additionally the class for a disassembling of a back end generated
variant section is introduced (no functionality yet) in case multiple
connectors provide data to the variant section.

loadFlexData now allows the following patterns:
* legacy response (which may be disassembled):
{
    changes: [],
    variantSection: { /* merged variant section */}
}
* plain single-layered response
{
    changes: [],
    variants: [],
    variantChanges: [],
    variantManagementChanges: [],
    variantDependentControlChanges: []
}
* a list of plain single layered-responses
[
    <plain single-layered response applied first>,
    <plain single-layered response applied second>,
    ...
]

Change-Id: I631d62f13b5c362acdee3fafdbeeaa21c19861e4