[FEATURE] Messaging: Message propagation for lists
With this change also messages for subentities are propagated to
OData V2 list bindings.

This change is a prerequisite for the table control to show the
number of errors for all enities within the table.

The new implementation supports OData V2 messages, but currently
no client-side validation messages.

Change-Id: I8b1accb39acbe58cc0e06b90bc69da486d84091a
JIRA: CPOUIFPHOENIXCORE-2840