[INTERNAL] sap.ui.model.odata.type.String: isDigitSequence & empty user input
If the user clears a field with digit sequence type and the type is not
nullable, it complained that the value must not be empty.

Now it has been decided that it shall set the value to "0" (poss. padded
to the max length) instead.

See the effects in
http://localhost:8080/testsuite/test-resources/sap/ui/core/demokit/sample/common/index.html?component=ViewTemplate.types
in the field named "DigitSequence (NUMC, nullable: false)"

Change-Id: I893e696dd7eaf2957ee9cca5fb1e21c61178b555
BCP: 1970359298
(cherry picked from commit 6976fd16bee01fad6a84c08f6669820073f8f9cf)