[INTERNAL] FocusHandler: focus during in-place DOM patching corrected
With new in-place DOM patching restoreFocus can be called with null
(e.g. when storePatchingControlFocusInfo does not update the patched
control focus info). As in such cases the LastControlFocusInfo was used
(which may not have been updated), it could lead to the issue that can
be seen in the linked incident.
This is now fixed by ensuring that last control focus info is updated
even when control focus info is NOT stored, ensuring any restoreFocus
calls (with null) still works with updated focus info from the control.

BCP: 1980333640
CR-Id: 002075125900004111372019
Change-Id: I5a6ecae27e7f0f395aa84c76db3785dde723525c