[INTERNAL][FIX] ui.Table: Scroll thumb min-height and min-width
When the ui.Table has big scrollable area (data sets of >~5000 rows) the
scroll thumb can disappear as its min-size is not limited by the
browser.
Defined min-height for the thumb in vertical scrollbar and min-width for
the thumb in horizontal scrollbar, to make sure that the scrollbar is
always visible.

Change-Id: I22cd17f29a6eaff68ba2888841596a28a793b203
BCP: 1980336282