[INTERNAL][FIX] sap.m.SelectDialog: Tap/onsapselect event delegates are now moved to the List items
Event delegates were on the List, instead of the List items,
and closing of the dialog was not preventable.

With this change, if someone has implemented some sort of custom list item,
they would be able to prevent closing of the Dialog and keep the old
behaviour- close the Dialog only on List item change.

BCP: 1980346450
Change-Id: Id12fcee55bb46e4fdf8a3814ff572b659a7453a9