[INTERNAL] ListBase: avoid binding parsing when grouping
When a column is grouped in the Responsive table which contains a value
like "12345><:{)*^", the ListBase creates a GroupHeaderListItem and sets
the title property in its constructor. The value is treated as a complex
binding by the ManagedObject and due to the "{" the binding parser fails
and throws an error in the console. So to fix the issue, we make use of
the setTitle method to apply the title to the GroupHeaderListItem. This
also ensures that such values are treated as plain text and not as
complex bindings.
BCP: 1980346380

Change-Id: I4cfc2430b2c01db4e54e02e171d52825ebab592c