[FIX] sap.m.Carousel: Changing pages with keys fast is now possible
If the arrow keys are pressed fast, now we check if animation has already
started. In such cases call the callbacks that would be usually called
when the animation completes and deregister listeners (this is already
possible with the mobify-carousel, so just call its methods).
Then just call .move, which will start another animation for the next element.

Fixes: https://github.com/SAP/openui5/issues/2653
Change-Id: I9fd03f3afc5efcd1a12a46bfc9c5d32f683ccf34
BCP: 1980357464