[INTERNAL] sap.m.Input: Sucess value state border is now 1 px wide
The Success semantic state is a validation/confirmation hint,
and should thus be easily distinguishable from semantic states such as e.g.
Warning and Error, which require more immediate attention and action from the user.
It should therefore use the same border width as the regular state.
This is valid for both Fiori 3 as well as Fiori 2.

Change-Id: I3bea6361479d1f307198bf8f43028be0308ddfcf
JIRA: BGSOFUIRODOPI-2116