[INTERNAL] ListBase: range selection fix
When performing range selection and the next selected item is the table
header row or the footer, then there was an exception logged in the console
because the next selectable item could not be found. To fix the issue, the
nextItem is first checked whether it is undefined, in this case the
selection is performed.
If the nextItem to select is not found then, the rangeSelection object is
cleared so that there is no unexpected selection action.
BCP: 1980394212

Change-Id: If34e74784718fee4ef9a4cb64c408264880eb29c