[FIX] sap.m.MultiComboBox: Sync selectedKeys & selectedItems
When MultiComboBox gets initialised setters of the keys and items
need to be handled as a special case. selectedkeys and selectdItems
in the inint of the control, should merge those fields.
While calling those setters individually should reset/replace the selection.

Therefore, calling setSelectedKeys in a row, should reset
the fields after the second call. The same should be applied for setSelectedItems.

BCP: 1980392874
Change-Id: I12c12f8112a2161f2e9f29bcca76c8450b6600a1