[FEATURE] StandardListItem: visual adaptation
Adapting the padding and height of the StandardListItem control
according
to the Visual specs.
The content that is to the left of the Thumbnail icon, i.e., selection
radio/checkbox will be top aligned and the content to the right of the
text, i.e., buttons (Detail and Delete) will be center aligned.
Additionally, the wrapping StandardListItem behavior is also adapted
according to the specs. Min-height also also added to the wrapping list
item.
Resolves:
JIRA: CPOUIFTEAMB-1389
Fixes: https://github.com/SAP/openui5/issues/2676

Change-Id: Ib515c84a036aaea696242ed8932a31b1a8d03efc