[FIX] sap.f.DynamicPage: Corrected pin button update
Background:
By design, if the DynamicPageHeader is bigger than a certain threshold,
it cannot be pinned

Problem:
If the header became bigger after rendering and the header
could not be snapped with scroll, the above check was skipped.
Snippix: 36516

Solution:
Ensured the check for header size is always made on child resize.

BCP: 1980353295

Change-Id: I3e774ab2b177b2888eddd1428db59f3c15c3a041