[FIX] sap.m.Text: Hyphenation for single words fixed in IE11
When wrapping type is set to "Hyphenated", css style "break-word: break-all"
breaks the hyphenation and makes it behave like normal word breaking.
In cases when hyphenation will be made, we should let it decide where
the words should break, and not add break-word at all.
The issue appread on IE11 only because there is style "word-break: break-word"
for webkit browsers. If this style is not there, the issue is visible on all
browsers.

Change-Id: Ic973345166031de8926d1541810c3616520dbb78
BCP: 1970355420