[INTERNAL] Commands: Move focusable span to static-ui-area
Yet, the focusable span element was added as last child
to the scoped control which defines the commands.
This lead to the issue, that the scroll position moves
unexpectedly when the span element isn't in the visible
content area of the browser.

With this change the focusable span element is added to the
static-ui-area and positioned absolute in the middle of the
screen.

Change-Id: I43faacb98a25a5bead9bbea522e89b298485aee7