[INTERNAL] ListItemBase: focus outline fix for highlight
When the highlight is set on the item, the focus outline of the item is
overlapped by the hightlight div. Hence, on focus of the item, the
hightlight div is position accordingly and width is reduced by 2px such
that the focus outline is visible on the all 4 sides for the list item.
Resolves:
BCP: 1980397581

Change-Id: I5c41b1c300f29a143865d6d784f19cd3dd63d213