[INTERNAL] sap/ui/fl/...CF-connectors merge compVariants into changes
with the new v1 apis from the personalization connector and the key user
connector the compVariants are sent in different properties to ease the
consumption on client side. Since this is not yet expected form the
client, it is merged directly in the connectors but a later change will
not result in an additional route to maintain on server side.

Change-Id: Iaea84a50b375712e90a905e797af3ac7c57315fe