[INTERNAL] sap.ui.integration.CardExplorer: Landing page adjusted
Landing page example adapted to use "ui-integration-card" element
and indentation of the code snippet is fixed.

Reasons for downport: This change fixes regression in the landing page
of the CardExplorer. It cannot affect any application, besides it.
The fix also prevents potential legal problem, because some images and
names were used as examples, which are not SAP branded and can bring
copyright issues.

Change-Id: I972266cd28fcc9f72ec9acaaddfb20f37787c400
JIRA: BGSOFUIRODOPI-2130
BCP: n/a
CR-Id: 002075125900004899992019