[INTERNAL] sap.m.PlanningCalendar: Visual fix in the row header
Changes in the StandardListItem's less caused visual issues in the
PlanningCalendar. Now our less has a stronger selector.
It's crucial this to be downported to 1.71 since without it the row
header when there's an image, the text which has to be right on its
right side, overlaps it.

Change-Id: I73cbeb4c4a014e2bd4470905a08e990c9ed4ed76
BCP: n/a
(cherry picked from commit d0eca9890abe7a679b8f2289b6cd036485abb20f)
CR-Id: 002075125900004968142019