[INTERNAL] sap.ui.fl.internal.apply.connectors.StaticFileConnector understands flexibility-bundle.json files
With the possibility of bundling fl variant & dependent objects into a single file it is mandatory for the StaticFileConnector to adjust.
In this change the Connector will check for a second file (instead only the changes-bundle.json).
The new file is called flexibility-bundle.json and has the same loading behavior as the changes-bundle.
The format of the file is expected to be the same as the exptected response of the Connector.loadFlexData call.

Change-Id: I38519b7901cc2dfa0a0059f982b13dbdac70f9f9
JIRA: UICSFLEX-6028