[INTERNAL] sap.ui.fl.internal.apply.connectors.StaticFileConnector understands flexibility-bundle.json files
INFO: This commit also removes the redundant parameter
"xx-flexBundleRequestForced" from the fl-coding. This parameter was
always set with the "componentPreload = off" and thus is obsolete. This
parameter is now checked directly instead.

With the possibility of bundling fl variant & dependent objects into a
single file it is mandatory for the StaticFileConnector to adjust.
In this change the Connector will check for a second file (instead only
the changes-bundle.json).
The new file is called flexibility-bundle.json and has the same loading
behavior as the changes-bundle.
The format of the file is expected to be the same as the expected
response of the Connector.loadFlexData call.

Change-Id: I990046175b2ae1767684a561f6e439581ea6ad2f
JIRA: UICSFLEX-6028