[INTERNAL] ui5loader: call sap.ui.require (sync) callbacks in micro task
In sync loading mode, sap.ui.require simulated its asynchronous API
contract by using a setTimeout(...,0). Together with the chaining of
sap.ui.require calls (as e.g. in the XMLTemplateProcessor), this could
result in a spec-imposed delay of 4ms per sap.ui.require call.

With this change, the callbacks are now executed in the next microtask,
not task.

Change-Id: I94db44464aff6d3df9d35c92bca099b26bd5e855