[INTERNAL][FIX] Demo Kit: Fixed overlapping in Avatar sample
Problem: At some point on XL grid size, the avatars at the bottom of
the sample are overlapping.

Root Cause: The grid configuration, of which the avatars are placed,
is incorrect.

Solution: On XL grid size now an avatar takes 33% of a row, instead 25%.
Now we have three avatars on a row, instead of four.
This allows more tolerance between them, hence no overlapping.

Additionally the avatars size is reduced from XL to L.

Change-Id: Id32977c992ff56887a381d1f07fdc5781782bd0b
BCP: 1970472626