[FIX] sap.m.OverflowToolbar: Empty Popover is not shown anymore
Problem: When a control has a Disappear OverflowToolbarPriority and
there is not enough space for it in the toolbar, the overflow button
appears but when it is clicked, an empty Popover is shown.

Solution: Check if there are other controls in the Popover with
different OverflowToolbarPriority and show the overflow button only when
needed.

Change-Id: I505d9c11ff01a97de04c72aecbefb3293a04a080
BCP: 1970480882