[INTERNAL] sap.m.PlanningCalendar: Visual fix in the row header
In the row header area there could be a thumbnail or an icon + text.
Before this fix the space between an icon and text was less than a
thumbnail and text. Now they are equal.
On top of that, in order to satisfy the VD spec, the width, height and
line-height of the icon in belize is also fixed.

Change-Id: I37c1a5716ad3b3b17ee7580e0ea062f4b426874d