[INTERNAL] ListItemBase: fix width for selection control
With the recent visual adaptation to the StandardListItem, the width of
the selection controls were set to 3rem. The checkbox control worked as
expected but the radio button rendered unexpected focus outline because
of the width, hence setting the width to 2rem in compact density as before
and applying a margin-right of 1rem to satisfy the visual appearence.
BCP: 1970476306

Change-Id: I40fe00602320c7f4ea2dd0f09df7e8d7ba683f4e