[INTERNAL] ColumnListItem: fix onfocusin event
When the focus is set on inner focusable controls within the ColumnListItem
and sticky feature is enabled on the Table, then the partially visible
control within the ColumnListItem should be entirely visible. The ColumnListItem
prevented this from happening and the sticky scroll handling was not triggered
within the ListItemBase which is fixed.
BCP: 1980340480

Change-Id: I53d2a8b6f0e58219f1647c31ef2b5017d0881987