[INTERNAL] Table: fix row border alignment issue
The th element had a style background-clip: padding-box which was
required in Edge browser as the focus outline of the item below the
sticky column headers was not visible on all sides. But this style
caused issue for other browsers where a gap was visible between the th
element's background color and bottom border when zoomed. To solve this
issue, the special style background-clip: padding-box is now applied
only for Edge browser.
BCP: 1970474570

Change-Id: I1486d853df8ab269aa58ff6630eef8317d308434